+++ src/main/java/org/sbrubbles/conditio/HandlerNotFoundException.java
@@ -5,7 +5,7 @@ import org.sbrubbles.conditio.policies.Policies;
 /**
  * Thrown when no working handler for a given condition was found. This may happen either when no available handler
  * could {@linkplain Handler#test(Object) handle} the condition, or when the ones that could opted to
- * {@linkplain Handler.Context#skip() skip} instead.
+ * {@linkplain Handler.Operations#skip() skip} instead.
  *
  * @see Handler
  * @see Handler.Context